Outputs derivatives in the available symbols.
Also introduces this horrible hack where commas are replaced with semicolons inside the diff function because otherwise the parser that ingests the available symbols chokes. This makes the whole thing a bit weird, as I designed the derivative token to be the same as the sympy function.

Possible fix: choose another separator for the list of available symbols. Problem: someone would need to go through the entire content repo and change commas to whatever else we choose.