Close the dropdown menu when "Sign up" is clicked
Previously clicking signup would only redirect you to the register
page and the login dropdown would remain open.
On small screens the dropdown could cover most of the page so it
would not be obvious you have moved to the register page.

Now the dropdown is hidden when the sign up button is clicked.